# More performance tweaks